target group inprovements/preprocessing_service/requirements.txt
flask==3.1.1
flask-restx==1.2.0
pyspark==3.5.5
numpy==2.2.4
torch --index-url https://download.pytorch.org/whl/cu126
pillow==11.0.0
opencv-python==4.11.0.86
google-cloud-storage==3.1.0